+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/internal/KotlinPsiFileFactory.kt
@@ -1,5 +1,8 @@
 package com.pinterest.ktlint.core.internal
 
+import com.pinterest.ktlint.core.KtLint
+import java.nio.file.Files
+import java.nio.file.Path
 import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.cli.jvm.compiler.EnvironmentConfigFiles
@@ -9,6 +12,7 @@ import org.jetbrains.kotlin.com.intellij.openapi.diagnostic.DefaultLogger
 import org.jetbrains.kotlin.com.intellij.openapi.diagnostic.Logger as DiagnosticLogger
 import org.jetbrains.kotlin.com.intellij.openapi.extensions.ExtensionPoint
 import org.jetbrains.kotlin.com.intellij.openapi.extensions.Extensions.getRootArea
+import org.jetbrains.kotlin.com.intellij.openapi.util.Disposer
 import org.jetbrains.kotlin.com.intellij.openapi.util.UserDataHolderBase
 import org.jetbrains.kotlin.com.intellij.pom.PomModel
 import org.jetbrains.kotlin.com.intellij.pom.PomModelAspect
@@ -20,17 +24,40 @@ import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.TreeCopyHandler
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import sun.reflect.ReflectionFactory
 
+internal class KotlinPsiFileFactoryProvider {
+    private lateinit var psiFileFactory: PsiFileFactory
+
+    @Synchronized
+    fun getKotlinPsiFileFactory(isFromCli: Boolean): PsiFileFactory =
+        if (::psiFileFactory.isInitialized) {
+            psiFileFactory
+        } else {
+            initPsiFileFactory(isFromCli).also { psiFileFactory = it }
+        }
+}
+
 /**
  * Initialize Kotlin Lexer.
  */
-internal fun initPsiFileFactory(): PsiFileFactory {
+internal fun initPsiFileFactory(isFromCli: Boolean): PsiFileFactory {
     DiagnosticLogger.setFactory(LoggerFactory::class.java)
 
     val compilerConfiguration = CompilerConfiguration()
     compilerConfiguration.put(CLIConfigurationKeys.MESSAGE_COLLECTOR_KEY, MessageCollector.NONE)
+    // Special workaround on JDK 1.8 when KtLint is used from shipped CLI
+    // to prevent Kotlin compiler initialization error
+    if (isFromCli && System.getProperty("java.specification.version") == "1.8") {
+        val extensionPath = extractCompilerExtension()
+        compilerConfiguration.put(
+            CLIConfigurationKeys.INTELLIJ_PLUGIN_ROOT,
+            extensionPath.toAbsolutePath().toString()
+        )
+    }
+
+    val disposable = Disposer.newDisposable()
 
     val project = KotlinCoreEnvironment.createForProduction(
-        {},
+        disposable,
         compilerConfiguration,
         EnvironmentConfigFiles.JVM_CONFIG_FILES
     ).project as MockProject
@@ -38,6 +65,25 @@ internal fun initPsiFileFactory(): PsiFileFactory {
     project.enableASTMutations()
 
     return PsiFileFactory.getInstance(project)
+}
+
+/**
+ * Note: this only works in CLI shadowed jar! 'extensions/compiler.xml' is absent in non-shadowed jar.
+ */
+private fun extractCompilerExtension(): Path {
+    KtLint::class.java.getResourceAsStream("/META-INF/extensions/compiler.xml").use { input ->
+        val tempDir = Files.createTempDirectory("ktlint")
+        tempDir.toFile().deleteOnExit()
+
+        val extensionsDir = tempDir.resolve("META-INF/extensions").also {
+            Files.createDirectories(it)
+        }
+        extensionsDir.resolve("compiler.xml").toFile().outputStream().buffered().use {
+            input!!.copyTo(it)
+        }
+
+        return tempDir
+    }
 }
 
 /**