+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-## [TO BE DETERMINED]
+## Unreleased
 
 ### Fixed
 - ClassCastException using ktlintFormat on class with KDoc in version 0.43.0 (`no-trailing-spaces`) ([#1270](https://github.com/pinterest/ktlint/issues/1270)