+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Fixed
+- Fix false positive indentation (`parameter-list-wrapping`) ([#1267](https://github.com/pinterest/ktlint/issues/1267))
+- Fix false positive indentation (`parameter-list-wrapping`) ([#1119](https://github.com/pinterest/ktlint/issues/1119))
+- Fix false positive indentation (`argument-list-wrapping`) ([#1045](https://github.com/pinterest/ktlint/issues/1045))
 
 ### Changed