+++ CHANGELOG.md
@@ -7,8 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Fixed
+- KtLint CLI 0.43 doesn't work with JDK 1.8 ([#1271](https://github.com/pinterest/ktlint/issues/1271))
 
 ### Changed
+- Update Kotlin version to `1.6.0` release
 
 ### Removed