+++ CHANGELOG.md
@@ -8,7 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 - KtLint CLI 0.43 doesn't work with JDK 1.8 ([#1271](https://github.com/pinterest/ktlint/issues/1271))
-
+- Fix false positive in rule spacing-between-declarations-with-annotations ([#1281] (https://github.com/pinterest/ktlint/issues/1281))
+- 
 ### Changed
 - Update Kotlin version to `1.6.0` release